chore: 각종 검색 쿼리 search feature로 이동
