+++ src/feature/Genre/graphql/query.ts
@@ -1,19 +1,5 @@
 import { graphql } from "$lib/graphql/__generated__"
 
-export const searchGenres_QUERY = graphql(/* GraphQL */ `
-    query searchGenres($word: String!) {
-        genresCollection(
-            filter: { name: { ilike: $word } }
-        ) {
-            edges {
-                node {
-                    ...GenresInfo
-                }
-            }
-        }
-    }
-`)
-
 export const getGenreInfoById_QUERY = graphql(/* GraphQL */ `
     query getGenreInfoById($id: UUID!) {
         genresCollection(
