+++ src/feature/Profile/graphql/index.ts
@@ -1,2 +1,1 @@
 export * from './mutation'
-export * from './query'
