+++ src/feature/search/graphql/index.ts
@@ -1,0 +1,1 @@
+export * from './query'
